Add :force_ssl option for forcing https
I ran into a use-case where I needed to redirect the following to the
(https://www) version of the url:
http://example.com
https://example.com
http://www.example.com
https://www.example.com

In the first case this would take two redirects, which is not the
end of the world. However, I figured since I was already using this
gem I'd check for an option I could pass that would ensure https was
used while it was ensuring the desired host. Since there wasn't one,
I built it out (including specs). It's a very minor option and is
backwards compatible (doesn't force a redirect if the option is
missing or if we're already using https).