Support temporary (302) redirects
With no explicit cache expiry set, a browser may choose to cache a 301
redirect for as long as it likes. This can cause problems if you choose
to alter where a domain redirects to in the future; browsers which cache
the request will continue to point to the old redirect.

By specifying the `temporary` option, a user can ask that
`Rack::CanonicalHost` return 302 redirects instead, causing the browser
not to cache the request, and making it easier to redirect the domain
somewhere else at a later date.

Usage:

```ruby
use Rack::CanonicalHost, 'example.com', temporary: true
```